Bug 2015410 - Part 1: Fix clang-tidy warnings in js/src/builtin/intl. r=spidermonkey-reviewers,arai
Fixes various clang-tidy warnings, like:
- `auto` instead of `auto`.
- Unnecessary explicit initialisation of class members.
- Missing default initialisation of class members.
- `length() == 0` instead of `empty()` tests.
- Missing explicit casts to narrow int types.
- Missing `default` case in `switch` statements.
- `bool` return statements which can be simplified.

Differential Revision: https://phabricator.services.mozilla.com/D282344

+++ js/src/builtin/intl/DurationFormat.cpp
@@ -684,7 +684,8 @@ static bool InitializeDurationFormat(
                          &fractionalDigits)) {
       return false;
     }
-    dfOptions->fractionalDigits = fractionalDigits.valueOr(-1);
+    dfOptions->fractionalDigits =
+        static_cast<int8_t>(fractionalDigits.valueOr(-1));
   }
   durationFormat->setOptions(dfOptions.release());
   AddCellMemory(durationFormat, sizeof(DurationFormatOptions),
